requirements.txt
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-api-python-client==2.184.0
requests==2.32.5
beautifulsoup4==4.14.2